fix: initialize project
