--- CHANGELOG.md
@@ -1,6 +1,0 @@
-# @lit/lit-starter-ts
-
-## 1.0.0
-
-### Major Changes
-- Initialize project with lit.dev
