--- dev/README.md
@@ -1,1 +1,0 @@
-This directory contains HTML files containing your element for development. By running `npm run build:watch` and `npm run serve` you can edit and see changes without bundling.
