--- dev/index.html
@@ -1,22 +1,0 @@
-<!doctype html>
-
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Demo</title>
-    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script src="../node_modules/lit/polyfill-support.js"></script>
-    <script type="module" src="../build/off-webcomponents.js"></script>
-    <style>
-      p {
-        border: solid 1px blue;
-        padding: 8px;
-      }
-    </style>
-  </head>
-  <body>
-    <off-webcomponents>
-      <p>This is child content</p>
-    </off-webcomponents>
-  </body>
-</html>
