--- index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Lit Starter Kit</title>
-  </head>
-  <body>
-    <a href="/dev/index.html">Component Demo</a>
-  </body>
-</html>
