--- src/test/off-webcomponents_test.ts
@@ -1,63 +1,0 @@
-
-/**
- * @license
- * Copyright 2021 Google LLC
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-import {OffWebcomponents} from '../off-webcomponents.js';
-
-import {fixture, assert} from '@open-wc/testing';
-import {html} from 'lit/static-html.js';
-
-suite('off-webcomponents', () => {
-  test('is defined', () => {
-    const el = document.createElement('off-webcomponents');
-    assert.instanceOf(el, OffWebcomponents);
-  });
-
-  test('renders with default values', async () => {
-    const el = await fixture(html`<off-webcomponents></off-webcomponents>`);
-    assert.shadowDom.equal(
-      el,
-      `
-      <h1>Hello, World!</h1>
-      <button part="button">Click Count: 0</button>
-      <slot></slot>
-    `
-    );
-  });
-
-  test('renders with a set name', async () => {
-    const el = await fixture(html`<off-webcomponents name="Test"></off-webcomponents>`);
-    assert.shadowDom.equal(
-      el,
-      `
-      <h1>Hello, Test!</h1>
-      <button part="button">Click Count: 0</button>
-      <slot></slot>
-    `
-    );
-  });
-
-  test('handles a click', async () => {
-    const el = (await fixture(html`<off-webcomponents></off-webcomponents>`)) as OffWebcomponents;
-    const button = el.shadowRoot!.querySelector('button')!;
-    button.click();
-    await el.updateComplete;
-    assert.shadowDom.equal(
-      el,
-      `
-      <h1>Hello, World!</h1>
-      <button part="button">Click Count: 1</button>
-      <slot></slot>
-    `
-    );
-  });
-
-  test('styling applied', async () => {
-    const el = (await fixture(html`<off-webcomponents></off-webcomponents>`)) as OffWebcomponents;
-    await el.updateComplete;
-    assert.equal(getComputedStyle(el).paddingTop, '16px');
-  });
-});
