+++ web-components/index.html
@@ -1,0 +1,23 @@
+<!doctype html>
+
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>Demo</title>
+    <script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+    <script src="node_modules/lit/polyfill-support.js"></script>
+    <script type="module" src="build/off-webcomponents.js"></script>
+    <style>
+      p {
+        border: solid 1px blue;
+        padding: 8px;
+      }
+    </style>
+  </head>
+  <body>
+    <h1>List of all available components</h1>
+    <div>
+      <hunger-question question="test ?" product='{"imgUrl": "https://upload.wikimedia.org/wikipedia/commons/thumb/9/99/Open_Food_Facts_logo_vertical_2022.svg/1200px-Open_Food_Facts_logo_vertical_2022.svg.png"}'></hunger-question>
+    </div>
+  </body>
+</html>
