+++ web-components/src/off-webcomponents.ts
@@ -1,0 +1,1 @@
+export { HungerQuestion } from "./hunger-question";
