+++ web-components/src/types.ts
@@ -1,0 +1,6 @@
+export type Product = {
+  name: string;
+  price: number;
+  description: string;
+  imgUrl: string;
+};
